rename files, add first text, load data from python
