+++ hexagon-layer/src/index.js
@@ -4,7 +4,9 @@ import './index.css';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
 
-const data = [];
+const data = window.coords;
+console.log(data);
+console.log("here it is!!")
 
 const MAPBOX_TOKEN = 'pk.eyJ1Ijoiam9hbm5lY2hlbmciLCJhIjoiYV9YSTdaZyJ9.tOevZpArPItdszzQl_GLJQ';
